changed to long

+++ DropperEquipment/src/main.cpp
@@ -45,7 +45,7 @@ bool slowStart = true;
 bool readyToStart = false;
 bool productionRun = false;
 volatile boolean turnDetected;
-double rotaryPosition;
+long rotaryPosition;
 int previousPosition;
 int stepsToTake;
 bool s1 = false;
@@ -66,7 +66,7 @@ int encoderCurrentState;
 int encoderPreviousState;
 int encoderCount = 0;
 
-double calculateDegrees(double rotaryPosition) //converts the steps the stepper has stepped to degrees //a 400 step goes 0.9 degrees per step. 200 stepper motor is 1.8 degrees per step. Currently 800!
+double calculateDegrees(long rotaryPosition) //converts the steps the stepper has stepped to degrees //a 400 step goes 0.9 degrees per step. 200 stepper motor is 1.8 degrees per step. Currently 800!
 {
   double result = rotaryPosition * .45; 
   return result;
